move main assembly dependent editor scripts into a new assembly
