+++ Assets/Scripts/Editor/EditorStandalone.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3aab38553094d694b9a0f948602e605d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
