+++ Assets/Scripts/Editor/EditorStandalone/AnyUnityObjectUIPropertyDrawer.cs
@@ -4,7 +4,7 @@ using UnityEditor;
 using UnityEngine;
 using Object = UnityEngine.Object;
 
-namespace Cosmobot.Editor
+namespace Cosmobot.Editor.Standalone
 {
     [CustomPropertyDrawer(typeof(AnyUnityObjectUIAttribute))]
     public class AnyUnityObjectUIPropertyDrawer : PropertyDrawer
