+++ Assets/Scripts/Editor/EditorStandalone/Cosmobot.Editor.Independent.asmdef.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 6ffbc5d30f071cc27b1cac04d9d631b0
-folderAsset: yes
-DefaultImporter:
+guid: a9e2552392a376b46aeb70595870ffb2
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
