+++ Assets/Scripts/Editor/ItemSystem.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e942b4ba5760b2a4b8b1f39587a76c77
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
