+++ Assets/Scripts/Editor/ItemSystem/ItemInfoInspector.cs.meta
